ERROR state: Add NVRs from the second set if available
Currently the output states there are no builds in the second set even
if there are.  The ERROR state signals the requested package is not
available in the first set, which is still correct.  This makes it
different from EXTRA.

However, None/None output is fairly misleading.  This should make the
situation fairly more obvious.

Signed-off-by: Petr Šabata <contyk@redhat.com>