[WIP] Script to rebuild the group of builds
Inputs are sidetag or list of NVRs.

Script takes the input and creates a list of packages which need to be
rebuilt for ELN. It uses getBuild for each package to get the source url, and then triggers rebuilds of the entire list.

Related: https://github.com/fedora-eln/eln/issues/2